Fixed ordering of habits Fixed completion of habits that should be completed multiple times
